Added gatsby navigate to get category buttons working
Started using a simple href material ui button but found that it was slow trying to load the next whole page while gatsby navigate was alot faster. Added test to see category selection is used on the search page. If it is, it will run a function given via props to change the search result category